bootstrap-template.yml: Install dbus as workaround
dbus is needed to set the hostname, but is currently not provided by the
role Oefenweb.hostname. As a workaround, we make sure that dbus is
installed before running the role.

While at it dropped the group_vars/all file which just holds the
dns_server variable which can be set in the bootstrap.yml and
bootstrap-template.yml playbooks.