Change upstream of paulfantom/ansible-restic
The upstream project paulfantom/ansible-restic has been archived.

There is also a problme where the role fails in check_mode
(if /tmp/restic-0.9.6-SHA256SUMS is missing, see:
https://github.com/paulfantom/ansible-restic/issues/66) but this issue
is not going to be fixed in upstream.

For now use fix in my fork. If it turns out to work well, it could be
merged to the master branch and later the role could be renamed to
jkirk.restic.