requirements.yml: Update donat-b.restic-rest
jkirk/ansible-restic-rest@0588f36 (from branch start-on-boot-2) is
definitely out of date.

We still have the problem that restic-rest-server.target is not started
on boot, which has the effect, that this role starts the target and
triggers a change although the restic-rest-server services are already
started. So we go back to upstream and try not diverge for now.