superlinter.yml: Do not ansible-lint ansible roles used in this template
The Ansible roles which are used in the site-* examples contain some
ansible-lint errors which can not be handled in this template.

We do not want to disable ansible-lint completely because the site-*
playbooks should be checked. To make this work, "${HOME}/.cache" needs
to be in exclude_paths in '.ansible-lint.yml'. '.cache' doesn't seem to be enough
(anymore).
The correct path for the ansible-lint configuration file for Ansible
Workflows is .github/linters/.ansible-lint.yml. Moved the configuration
file.

Without FILTER_REGEX_EXCLUDE the files in .github also get syntax
checked by ansible-lint and such errors are shown:

  WARNING  Listing 1 violation(s) that are fatal
  .github/linters/.ansible-lint.yml:18:1: syntax-check A playbook must be a list of plays, got a <class 'ansible.parsing.yaml.objects.AnsibleMapping'> instead
  Error: syntax-check A playbook must be a list of plays, got a <class 'ansible.parsing.yaml.objects.AnsibleMapping'> instead

We also switched to the slim image super-linter/slim@v4 as it uses up
much less resources, see: https://github.com/marketplace/actions/super-linter#slim-image