Enable fact caching
The default cache plugin is the memory plugin, which only caches the
data for the current execution of Ansible (of each play).

gathering controls the default policy of fact gathering.
The default is `implicit` where the cache plugin will be ignored and
facts will be gathered per play unless ‘gather_facts: False’ is set.

With `smart` each new host that has no facts discovered will be scanned,
but if the same host is addressed in multiple plays it will not be
contacted again in the run.

See: https://docs.ansible.com/ansible/latest/reference_appendices/config.html#default-gathering

The default cache plugin is `memory`.
With `jsonfile` the cache plugin will save the JSON files.

See:

* https://docs.ansible.com/ansible/latest/reference_appendices/config.html#cache-plugin
* https://docs.ansible.com/ansible/latest/collections/ansible/builtin/jsonfile_cache.html#ansible-collections-ansible-builtin-jsonfile-cache

The fact_caching_timeout defaults to 86000.
But if we want to set it to another value, put a commented out line for
convenience.